requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7
django-allauth==65.12.0
django-countries==7.6.1
django-phonenumber-field==8.3.0
django-widget-tweaks==1.5.0
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gunicorn==22.0.0
idna==3.10
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
propcache==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
stripe==7.0.0
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.7.0
xhtml2pdf==0.2.16
yarl==1.22.0
